requirements.txt
extract-msg
python-docx
python-pptx
pdfminer.six
chardet
lxml
pillow
streamlit
openai
gradio 